+++ Classes/Views/PBSourceViewBadge.m
@@ -115,7 +115,7 @@
 
 + (NSImage *) numericBadge:(NSInteger)number forCell:(NSTextFieldCell *)cell
 {
-	return [self badge:[NSString stringWithFormat:@"%d", number] forCell:cell];
+	return [self badge:[NSString stringWithFormat:@"%ld", (long)number] forCell:cell];
 }
 
 @end
